backend/pyproject.toml
[project]
name = "youtube-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest>=9.0.1",
    "strands-agents>=1.19.0",
    "strands-agents-tools>=0.2.17",
    "youtube-transcript-api>=1.2.3",
]
